Installing boto at provision time so tags_to_grains works.
We need boto installed because we use it in template deploy to convert
the AWS tags to salt grains. The version available as a package is not
the most recent (2.20 at the time of writing compared to 2.37) but it is
enough for this script to work. If we need a more recent version then we
can manage that via salt states later on.

To make this even easier bootstrap-salt has this option which we use:

>  -p  Extra-package to install while installing salt dependencies. One package
>      per -p flag. You're responsible for providing the proper package name.