Add a fab task to delete the tar file from upload_salt task.
As the doc string says, we need to delete that file (or rather empty the
bucket, but this is safter) before cfn_delete can run.

Expected usage:

   fab aws:... salt.delete_tar cfn_delete

The downside to this is that we delete the tar first, then we ask if you
want to delete the stack. For instance the output of the above command
tasks is:

    fab load_env salt.delete_tar cfn_delete
    [localhost] local: aws s3 --profile default rm s3://ashsalttest-ash-test-a568bfa1-salt/srv.tar.gpg
    delete: s3://ashsalttest-ash-test-a568bfa1-salt/srv.tar.gpg
    Are you really sure you want to blow away the whole stack!? (y/n)
    y

    STACK ashsalttest-ash-test-a568bfa1 DELETING...

Also if you forget to add this task the delete fails, so I will
investigate ways we could make this automatic when calling the cfn_delete task