Make delete_tar happen automatically in cfn_delete
We do this by wrapping cfn_delete like we do with cfn_create and then
passing the delete_tar function (not the task) in the
pre_delete_callbacks parameter which cfn_delete invokes for us after
confirming with the user that they want to delete the stack but before
actually calling Cloudformation#delete.

This needs https://github.com/ministryofjustice/bootstrap-cfn/pull/144
(which is likely to be 0.5.7) of bootstrap-cfn for this functionality to
work.