Fix error in creating salt tar when you project has a file in ./salt/_grains/
The problem was the shutil.copytree expects the target directory to not
exists, and if it already exists then will raise an OSError. This could
happen because we first copy the project's ./salt dir to the tmp dir
(which in this case had a ./salt/_grains/neighbours.py), and then we try
to copy the contrib/srv/salt/_grains/ dir which fails because the
previous step created that folder already.

The best solution I can found (as mentioned in
http://stackoverflow.com/a/12687372/439189) is to copy the
implementation of copytree from the shutil module and put a guard around
the the makedir