Fix log path in salt_utils script
This was a subtle one. If you pass file:///path/to/some/file. Salt
passes this into python's logging.handlers.SysLogHandler which will try
and connect to a Unix socket at this path. If you omit the file:// then
salt creates a regular log file at the path which is what we want.
Because the file we passed wasn't a socket we couldn't connect and this
actually causes an exception to be raised.