Fix KMS key retrieval
The AWS api call to ListAliases has a default results limit of
50 before it starts paginating results. If we dont check through
all the results we may miss out on finding an existing key, try
to create one during stack creation, and fail since it actually
does exist, just not on the first page of results. This change
makes sure we keep going through paginated results until get
to the end or find our key.