Update bootstrap wait check
While waiting for bootstraping saltstack to finish we first wait on
ssh access, then wait for the creation of the bootstrap_done file.
This procedure will break if we lose ssh access after creation of the
ec2 instances, which can happen when we reboot after security updates
during cloud-init.

This change will attempt to handle possibly losing ssh access and
recovering while waiting for saltstack bootstrapping to complete.

* Set bootstrap done file location globally as env
* Add logging
* Attempt to recover on waiting for bootstrap completion when we lose
ssh connection