Refactor salt_utils to synchronise remote data more reliably
The current 2014.7 version of SaltStack doesn't handle updating
or synchronising dynamic module data and then accessing it in a
consistent manner. This leads to to inconsistencies in the versions
of custom grains, states, etc, available at different points throughout
a set of salt operations loaded into memory. Note that later versions
of salt deal with this situation better with the addition of
reload_grains and similar new methods to update internal state to reflect
the updated dynamic data during the lifetime of a salt-minion.

When the salt python modules are loaded in the script, any sequential
or concurrent operations can get custom grain information in a range of
states from fully updated to all custom grain data missing. This is due
to the data being accessed in a range of different ways dependent on the
operation, from accessing the salt.loader which might hold fully updated
data, to accessing through __grains__ such as in the template parsing
and rendering of pillar data, which might not reflect any changes
since the modules were loaded. This is an understood consequence
of the way and order salt updates grains and modules, with some
mitigation in later salt versions.

Since we don't use master/minions setups, and hence are not running
salt-minion. This means that we do not have a problem with salts
expectation that grains are relatively static throughout the lifetime
of a minion, and only that they are static throughout the
lifetime of a salt call instantiation.

This change takes a simple approach to consistently downloading and
synchronising data while maintain back compatibility. The original
salt_utils script is seperated out into two distinct scripts, one to
download and synchronise salt config data, the second to apply the state.
Along with some additional caching and syncing to provide some robustness
in the active data state, his ensures that the data accessed by the salt
state called reflects that that is expected.

* Move update and state application functionality into distinct scripts
* Seperate out data synchronising code into a distinct functions and scripts
* Improve the quality of logging
* Allow specification of logging level on the command line
* Replace external subprocess calls to tar with in-code tarfile calls
* Fix pep8 warnings
* Handle syntax parsing of saltstack output to print errors better