Fix output type in github.get_paginated_content
The json response from the Github API (and by extension the `out`
variable in `get_paginated_content`) should be a `list` not a `dict`.

It looks like this typo was introduced in a refactor in commit cbca00e
where the out variable was wrongly assumed to be a `dict`. It was missed
by the tests because the `get_paginated_content` function was always
mocked. I added an additional test class to specfically test that
function.

Also, providing mutable values as function defaults causes unexpected
behaviour where the same mutable value is shared across all calls to the
function in that thread. I have added a test for this behaviour and
moved the default checking into the function.