Fix `setup.py tests`
The tests depend on nose; explicitly in `test_salt_utils_state.py` but
also implicitly otherwise the test scanner can't seem to find the tests.
It can if they're listed in the `tests/__init__.py` so I assume nose's
test scanner is used if available.

The salt version is updated as `2014.7.5` is no longer available in
pypi. I have updated to the oldest version available so that it's as
small a change as possible and I'm not confident enough in this library
to push it any further.

I did originally update the salt version in the `.travis.yml` but I
couldn't figure out how to stop flake8 failing.